Refactor dao to directory
