+++ services/api/src/dao/customer.js
@@ -1,0 +1,85 @@
+const R = require('ramda');
+const {
+  knex,
+  ifNotAdmin,
+  whereAnd,
+  inClause,
+  inClauseOr,
+  query,
+  prepare,
+} = require('./utils');
+
+const addCustomer = sqlClient => async (cred, input) => {
+  if (cred.role !== 'admin') {
+    throw new Error('Project creation unauthorized.');
+  }
+
+  const prep = prepare(
+    sqlClient,
+    `CALL CreateCustomer(
+        :id,
+        :name,
+        ${input.comment ? ':comment' : 'NULL'},
+        ${input.private_key ? ':private_key' : 'NULL'}
+      );
+    `,
+  );
+  const rows = await query(sqlClient, prep(input));
+  const customer = R.path([0, 0], rows);
+
+  return customer;
+};
+
+
+const getCustomerByProjectId = sqlClient => async (cred, pid) => {
+  const { customers, projects } = cred.permissions;
+  const str = `
+      SELECT
+        c.id,
+        c.name,
+        c.comment,
+        c.private_key,
+        c.created
+      FROM project p
+      JOIN customer c ON p.customer = c.id
+      WHERE p.id = :pid
+      ${ifNotAdmin(
+        cred.role,
+        `AND (${inClauseOr([['c.id', customers], ['p.id', projects]])})`,
+      )}
+    `;
+  const prep = prepare(sqlClient, str);
+
+  const rows = await query(sqlClient, prep({ pid }));
+
+  return rows ? rows[0] : null;
+};
+
+const deleteCustomer = sqlClient => async (cred, input) => {
+  if (cred.role !== 'admin') {
+    throw new Error('Unauthorized');
+  }
+  const prep = prepare(sqlClient, 'CALL deleteCustomer(:name)');
+
+  const rows = await query(sqlClient, prep(input));
+
+  // TODO: maybe check rows for changed values
+  return 'success';
+};
+
+const getAllCustomers = sqlClient => async (cred, args) => {
+  const where = whereAnd([
+    args.createdAfter ? 'created >= :createdAfter' : '',
+    ifNotAdmin(cred.role, `${inClause('id', cred.permissions.customers)}`),
+  ]);
+  const prep = prepare(sqlClient, `SELECT * FROM customer ${where}`);
+  const rows = await query(sqlClient, prep(args));
+  return rows;
+};
+
+module.exports = {
+  addCustomer,
+  deleteCustomer,
+  getAllCustomers,
+  getCustomerByProjectId,
+};
